ocrd_models/requirements.txt
ocrd_utils == 1.0.0a1
lxml
pyyaml
